+++ src/protontricks/data/scripts/bwrap_launcher.sh
@@ -1,0 +1,63 @@
+#!/bin/bash
+# Helper script
+set -o errexit
+
+BLACKLISTED_ROOT_DIRS=(
+    /bin /dev /lib /lib64 /proc /run /sys /var /usr
+)
+
+ADDITIONAL_MOUNT_DIRS=(
+    /run/media "$PROTON_PATH" "$WINEPREFIX"
+)
+
+mount_dirs=()
+
+# Add any root directories that are not blacklisted
+for dir in /* ; do
+    if [[ ! -d "$dir" ]]; then
+        continue
+    fi
+    if [[ " ${BLACKLISTED_ROOT_DIRS[*]} " =~ " $dir " ]]; then
+        continue
+    fi
+    mount_dirs+=("$dir")
+done
+
+# Add additional mount directories, including the Wine prefix and Proton
+# installation directory
+for dir in "${ADDITIONAL_MOUNT_DIRS[@]}"; do
+    if [[ ! -d "$dir" ]]; then
+        continue
+    fi
+
+    already_mounted=false
+    # Check if the additional mount directory is already covered by one
+    # of the existing root directories.
+    # Most of the time this is the case, but if the user has placed the Proton
+    # installation or prefix inside a blacklisted directory (eg. '/lib'),
+    # we'll want to ensure it's mounted even if we're not mounting the entire
+    # root directory.
+    for mount_dir in "${mount_dirs[@]}"; do
+        if [[ "$dir" =~ ^$mount_dir ]]; then
+            # This directory is already covered by one of the existing mount
+            # points
+            already_mounted=true
+            break
+        fi
+    done
+
+    if [[ "$already_mounted" = false ]]; then
+        mount_dirs+=("$dir")
+    fi
+done
+
+mount_params=()
+
+for mount in "${mount_dirs[@]}"; do
+    mount_params+=(--filesystem "${mount}")
+done
+
+
+exec "$STEAM_RUNTIME_PATH"/run --share-pid --launcher \
+"${mount_params[@]}" -- \
+--socket="$PROTONTRICKS_TEMP_PATH/launcher.sock"